Revert "Merge branch 'master' into expose_hotzones"
This reverts commit 8e6a9b8110953d9ad45ec6ba2a93f752f48eaa34, reversing
changes made to 8df9bf784c6e7d3227acd940688fb8fe8a289286.

+++ editor/plugins/gpu_particles_2d_editor_plugin.cpp
@@ -60,7 +60,7 @@ void GPUParticles2DEditorPlugin::_file_selected(const String &p_file) {
 void GPUParticles2DEditorPlugin::_menu_callback(int p_idx) {
 	switch (p_idx) {
 		case MENU_GENERATE_VISIBILITY_RECT: {
-			double gen_time = particles->get_lifetime();
+			float gen_time = particles->get_lifetime();
 			if (gen_time < 1.0) {
 				generate_seconds->set_value(1.0);
 			} else {
@@ -100,7 +100,7 @@ void GPUParticles2DEditorPlugin::_menu_callback(int p_idx) {
 }
 
 void GPUParticles2DEditorPlugin::_generate_visibility_rect() {
-	double time = generate_seconds->get_value();
+	float time = generate_seconds->get_value();
 
 	float running = 0.0;
